notification/requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.12.0
google-cloud-logging==3.8.0
google-api-python-client==2.105.0
python-dotenv==1.0.0